backend/requirements.txt
# backend/requirements.txt

aniso8601==9.0.1
attrs==24.2.0
azure-common==1.1.28
azure-core==1.30.2
azure-identity==1.17.1
azure-keyvault-secrets==4.8.0
azure-mgmt-core==1.4.0
azure-mgmt-cosmosdb==9.5.1
azure-mgmt-keyvault==10.3.1
blinker==1.8.2
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
dnspython==2.6.1
flasgger==0.9.7.1
Flask==3.0.3
Flask-Cors==4.0.1
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gunicorn==22.0.0
idna==3.7
iniconfig==2.0.0
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
mistune==3.0.2
msal==1.30.0
msal-extensions==1.2.0
packaging==24.1
pluggy==1.5.0
portalocker==2.10.1
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
pymongo==4.8.0
pytest==8.3.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
six==1.16.0
SQLAlchemy==2.0.32
typing_extensions==4.12.2
urllib3==2.2.2
waitress==3.0.0
Werkzeug==3.0.3
